+++ lib/src/android/webview_feature.dart
@@ -190,7 +190,11 @@ class WebViewFeature_ {
 
   ///This feature covers [UserScriptInjectionTime.AT_DOCUMENT_START].
   static const DOCUMENT_START_SCRIPT =
-      const AndroidWebViewFeature_._internal("DOCUMENT_START_SCRIPT");
+      const WebViewFeature_._internal("DOCUMENT_START_SCRIPT");
+
+  ///This feature covers [InAppWebViewSettings.willSuppressErrorPage].
+  static const SUPPRESS_ERROR_PAGE =
+      const WebViewFeature_._internal("SUPPRESS_ERROR_PAGE");
 
   ///Return whether a feature is supported at run-time. On devices running Android version `Build.VERSION_CODES.LOLLIPOP` and higher,
   ///this will check whether a feature is supported, depending on the combination of the desired feature, the Android version of device,
@@ -394,6 +398,10 @@ class AndroidWebViewFeature_ {
   ///This feature covers [UserScriptInjectionTime.AT_DOCUMENT_START].
   static const DOCUMENT_START_SCRIPT =
       const AndroidWebViewFeature_._internal("DOCUMENT_START_SCRIPT");
+
+  ///This feature covers [InAppWebViewSettings.willSuppressErrorPage].
+  static const SUPPRESS_ERROR_PAGE =
+    const AndroidWebViewFeature_._internal("SUPPRESS_ERROR_PAGE");
 
   ///Return whether a feature is supported at run-time. On devices running Android version `Build.VERSION_CODES.LOLLIPOP` and higher,
   ///this will check whether a feature is supported, depending on the combination of the desired feature, the Android version of device,